Fix wait timer for slideshows
xv can do a slideshow using the -wait <sec> option. However, when the time
is more than about 1 second, xv waits an extremely long time before going
to the next picture, making the option completely useless.

The reason for the problem is that the xv source uses the obsolete value
CLK_TCK to measure figure out how long it has to wait, and since it is
obsolete, it uses a fallback of CLOCKS_PER_SEC. However, this latter value
is something completely different (and always 1000000). Instead, it should
use sysconf(_SC_CLK_TCK).

http://bugzilla.rpmfusion.org/1929