Fix build with link-time optimization
When building with link-time optimization enabled, the compiler errors
out due to some arrays being declared with different sizes in different
compilation units. All of the problematic declarations are of character
arrays of size MAXPATHLEN in xv.h, which is included by all of xv's
source files.

However, the value of the token MAXPATHLEN varies depending on whether
or not the token NEEDSDIR is defined at the time of xv.h's inclusion. If
it is defined, MAXPATHLEN will be defined via system headers, typically
to a value of 4096. If it is not defined, a default value of 256 is
used. The NEEDSDIR mechanism was probably created to speed up
compilation by not including files unnecessarily on the very old and
slow systems that xv was originally written for, and is almost certainly
unnecessary now. Hence, the fix for this issue is to get rid of NEEDSDIR
altogether and just include the necessary files unconditionally.

Fixes #25.